organizes other services to individual modules
