+++ src/services/indicators/get_ma.py
@@ -4,7 +4,7 @@ from typing_extensions import OrderedDict
 from src.cache import cache
 from src.schemas.analysis import SMAData, Analysis, EMAData, WMAData, VWMAData, Indicator
 from src.schemas.time_series import TimePeriod, Interval
-from src.services.get_historical import get_historical_quotes
+from src.services.historical.get_historical import get_historical_quotes
 
 
 @cache(expire=60, market_closed_expire=600)
