+++ src/services/indicators/get_summary_analysis.py
@@ -6,7 +6,7 @@ from stock_indicators.indicators import get_ema, get_wma, get_vwma, get_rsi, get
 from src.cache import cache
 from src.schemas.analysis import SummaryAnalysis, AROONData, BBANDSData, SuperTrendData, IchimokuData, MACDData
 from src.schemas.time_series import Interval, TimePeriod
-from src.services.get_historical import get_historical_quotes
+from src.services.historical.get_historical import get_historical_quotes
 
 
 async def get_summary_sma(quotes, periods, sma=None) -> list:
